+++ app/client/src/pages/setup/SignupSuccess.tsx
@@ -30,9 +30,8 @@ export function SignupSuccess() {
   const showStarterTemplatesInsteadofBlankCanvas = useFeatureFlag(
     FEATURE_FLAG.ab_show_templates_instead_of_blank_canvas_enabled,
   );
-  const isEnabledForCreateNew = useFeatureFlag(
-    FEATURE_FLAG.ab_create_new_apps_enabled,
-  );
+  const isEnabledForCreateNew = true;
+  useFeatureFlag(FEATURE_FLAG.ab_create_new_apps_enabled);
 
   useEffect(() => {
     PerformanceTracker.stopTracking(PerformanceTransactionName.SIGN_UP);
